Corrected some database fields : "strength" was misspelled "strenght" caused by database schema "gold" was marked as "point"
Data migrations :
put some values in "Camp", "Level" and "Characters" tables to correctly initialise contextual information. Need to supress database and launch migration to update (in dev. not in production)
Character is only added to let user perform battle even if there is no player at an affordable level.

Templating :
Sended opponent camp "Characters" info to "opponent" with render() method and applied it to the template with Jinja